YT-21844: Cypress node reachability
Previously, Cypress nodes were only partially aware of their location in the global tree. These changes upgrade Cypress nodes' behavior by introducing a concept of reachability. A node, either trunk or branched, is considered to be reachable iff it can be resolved from the root. This is a (branched) subtree's property, that is affected by changes to the tree's topology. The latter occurs mainly through altering map nodes (attach/detach child, merge). Portal node is reachable by default.
In a way, each transaction defines a separate layer of node reachability, with a complex relation between nested layers. For example, a branch node may be reachable, while the originating node isn't due to removal of the originating node's ancestor, but keeping it alive in the nested transaction. A merge call resolves such conflicts, keeping the integrity of the global tree.
64b051ed4fbb14da5060f9f9d6ee744f3b65af82